requirements.txt
discord==0.16.12
python-aiml==0.9.1
asyncio